move stats from valence to stats_utils
